UI overhaul and admin interface start
