--- onto_spread_ed/routes/release.py
@@ -1,10 +1,0 @@
-import functools
-
-from flask import (
-    Blueprint, flash, g, redirect, render_template, request, session, url_for
-)
-from werkzeug.security import check_password_hash, generate_password_hash
-
-bp = Blueprint('release', __name__, url_prefix='/release')
-
-# def
