+++ onto_spread_ed/templates/admin/dashboard.html
@@ -1,0 +1,43 @@
+{% extends "base.html" %}
+{% block title %}{{ config['APP_TITLE'] }}{% endblock %}
+{% block head %}
+    {{ super() }}
+{% endblock %}
+{% block content %}
+
+    <div class="row align-stretch">
+        <div class="col-sm-6 col-md-4 col-lg-3 m-2">
+            <a class="card bg-light" href="/admin/release"
+               style="height: 100%; text-decoration: none; border-width: 2px 2px 4px 2px">
+                <div class="section p-5 text-center fs-2 bg-light">
+                    <i class="fas fa-upload"></i>
+                </div>
+
+                <div class="card-body">
+                    <h4>Release</h4>
+                    <p>
+                        Start the release process and release a new version of ontologies
+                    </p>
+                </div>
+            </a>
+        </div>
+        <div class="col-sm-6 col-md-4 col-lg-3 m-2">
+
+            <a class="card bg-light" href="/admin/rebuild-index"
+               style="height: 100%; text-decoration: none; border-width: 2px 2px 4px 2px">
+                <div class="section p-5 text-center fs-2 bg-light">
+                    <i class="fas fa-folder-open"></i>
+                </div>
+
+                <div class="card-body">
+                    <h4>Rebuild index</h4>
+                    <p>
+                        Force a complete rebuild of the index to ensure the search finds all entries.
+                    </p>
+                </div>
+            </a>
+        </div>
+    </div>
+
+
+{% endblock %}
